Initial OSX support. Lots of stuff missing like any working GS/SPU plugin. WIP tagged in code 'TODO OSX'

+++ common/src/Utilities/AlignedMalloc.cpp
@@ -62,7 +62,7 @@ void* __fastcall pcsx2_aligned_realloc(void* handle, size_t size, size_t align)
 		free( handle );
 #else
 		AlignedMallocHeader* header = (AlignedMallocHeader*)((uptr)handle - headsize);
-		memcpy_fast( newbuf, handle, std::min( size, header->size ) );
+		memcpy_fast( newbuf, handle, std::min( size, (size_t)header->size ) );
 		free( header->baseptr );
 #endif
 	}
